requirements.txt
kivy==2.1.0
plyer==2.1.0
pyjnius==1.5.0
android==1.0